% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{ler_morador}
\alias{ler_morador}
\alias{ler_desp_col}
\alias{ler_cad_col}
\alias{ler_desp_ind}
\alias{ler_aluguel}
\alias{ler_rend_trab}
\alias{ler_rend_outros}
\alias{ler_domicilio}
\title{Ler tabelas da POF}
\usage{
ler_morador(ano)

ler_desp_col(ano)

ler_cad_col(ano)

ler_desp_ind(ano)

ler_aluguel(ano)

ler_rend_trab(ano)

ler_rend_outros(ano)

ler_domicilio(ano)
}
\arguments{
\item{ano}{Ano da pesquisa}
}
\description{
Ler tabelas da POF
}
