% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{ler_pof}
\alias{ler_pof}
\title{Utilidade para ler arquivos de microdados}
\usage{
ler_pof(arquivo, tamanhos, nomes)
}
\arguments{
\item{arquivo}{caminho para arquivo}

\item{tamanhos}{tamanho das colunas}

\item{nomes}{nomes das colunas}
}
\description{
Utilidade para ler arquivos de microdados
}
